% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldiff.R
\name{alldiff}
\alias{alldiff}
\title{string representing an all_different constraint over the variables in the
vector p}
\usage{
alldiff(p)
}
\arguments{
\item{p}{}
}
\description{
string representing an all_different constraint over the variables in the
vector p
}
