% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minion.R
\name{discrete_matrix}
\alias{discrete_matrix}
\title{Discrete Matrix}
\usage{
discrete_matrix(nrow, ncol, domain, name = "l")
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{domain}{Domain of variables}

\item{name}{Matrix name}
}
\value{
A dataframe representing a nrow x ncol matrix
over the given domain.
}
\description{
Create a dataframe of discrete variables.
}
