% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minion.R
\name{minion_output}
\alias{minion_output}
\title{Minion Output}
\usage{
minion_output(x)
}
\arguments{
\item{x}{List with variables, search, unary_constraints,
binary_constraints and binary_vec_constraints}
}
\value{
A complete Minion 3 input string.
}
\description{
Given a list with members variables, search, unary_constraints,
binary_constraints and binary_vec_constraints return the Minion
3 input string for the defined constraint program.
}
